fix: redo oAuth to handle mobile devices
