+++ frontend/src/App/App.js
@@ -2,12 +2,13 @@ import './App.css';
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import { ThemeProvider } from '@mui/material';
-import Nav from '../components/NavBar/Nav';
+import Nav from '../components/Nav';
 import Homepage from '../routes/homepage';
 import Footer from '../components/footer';
 import LostPage from '../routes/lostPage';
 import FoundPage from '../routes/foundPage';
-import Form from '../components/Form';
+import GoogleLogin from '../routes/GoogleLogin';
+import Form from '../routes/Form';
 import theme from '../utils/AppTheme';
 import { UserContextProvider } from '../utils/UserContext';
 import { LostItemsContextProvider } from '../utils/LostItemsContext';
@@ -26,6 +27,7 @@ function App() {
                                 <Route exact path="/lost" element={<LostPage />} />
                                 <Route exact path="/found" element={<FoundPage />} />
                                 <Route exact path="/report/form" element={<Form />} />
+                              <Route exact path="/google" element={<GoogleLogin />} />
                             </Routes>
                             <Footer />
                         </div>
