+++ frontend/src/index.js
@@ -5,9 +5,7 @@ import App from './App/App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
+    <App />,
 );
 
 // If you want to start measuring performance in your app, pass a function
